requirements.txt
streamlit
PyPDF2~=3.0.1
fpdf~=1.7.2
python-docx
